ai-agent-service/requirements.txt
fastapi>=0.116.0
httpx>=0.28.0
openai>=1.99.0
pydantic>=2.11.0
python-dotenv>=1.1.0
uvicorn>=0.35.0
